Atom* types
- Fix long
- Remove 'InPlace' methods
- Standardise Snapshot() method

Restore 32bit Length property so that pattern-matching works (yeah, really)

Restore 32bit indexer property so that pattern-matching works (yeah, really)

+++ LanguageExt.Core/Concurrency/VectorClock/VectorClock.A.cs
@@ -100,7 +100,7 @@ public record VectorClock<A>(Seq<(A, long)> Entries)
     /// <summary>
     /// The number of entries in the vector clock.
     /// </summary>
-    public int Count =>
+    public long Count =>
         Entries.Count;
 
     /// <summary>
